Move core plugin registration to klh and out of session.
Fix some bad error handling on KlhClient.
